[auth, marketplace] Move marketplace/utils/auth into auth/utils/userAuthenticated.
