+++ app/src/app/index.jsx
@@ -60,7 +60,7 @@ import DashboardEditor from '$editor/dashboard'
 import ModalRoot from '$shared/components/ModalRoot'
 import Notifications from '$shared/components/Notifications'
 import { formatPath } from '$shared/utils/url'
-import { userIsAuthenticated } from '$mp/utils/auth'
+import { userIsAuthenticated } from '$auth/utils/userAuthenticated'
 import links from '../links'
 import history from '../history'
 import '../analytics'
