+++ app/src/auth/components/ForgotPasswordPage/index.jsx
@@ -2,7 +2,7 @@
 
 import React, { useContext, useCallback } from 'react'
 import { I18n, Translate } from 'react-redux-i18n'
-import { userIsNotAuthenticated } from '$mp/utils/auth'
+import { userIsNotAuthenticated } from '$auth/utils/userAuthenticated'
 
 import AuthFormProvider from '../AuthFormProvider'
 import AuthFormContext from '../../contexts/AuthForm'
