+++ app/src/auth/components/RegisterPage/index.jsx
@@ -4,7 +4,8 @@ import React, { useCallback, useContext } from 'react'
 import qs from 'query-string'
 import * as yup from 'yup'
 import { I18n, Translate } from 'react-redux-i18n'
-import { userIsNotAuthenticated } from '$mp/utils/auth'
+import { userIsNotAuthenticated } from '$auth/utils/userAuthenticated'
+import getSessionToken from '$auth/utils/getSessionToken'
 
 import useIsMountedRef from '$shared/hooks/useIsMountedRef'
 import useOnMount from '$shared/hooks/useOnMount'
@@ -20,7 +21,6 @@ import Checkbox from '../Checkbox'
 import AuthStep from '../AuthStep'
 import AuthLayout from '../AuthLayout'
 
-import getSessionToken from '$auth/utils/getSessionToken'
 import post from '../../utils/post'
 import onInputChange from '../../utils/onInputChange'
 import schemas from '../../schemas/register'
