+++ app/src/auth/components/ResetPasswordPage/index.jsx
@@ -9,9 +9,9 @@ import { push } from 'connected-react-router'
 
 import useIsMountedRef from '$shared/hooks/useIsMountedRef'
 import useOnMount from '$shared/hooks/useOnMount'
-import AuthFormProvider from '../AuthFormProvider'
-import { userIsNotAuthenticated } from '$mp/utils/auth'
+import { userIsNotAuthenticated } from '$auth/utils/userAuthenticated'
 import AuthFormContext from '$auth/contexts/AuthForm'
+import AuthFormProvider from '../AuthFormProvider'
 import AuthPanel from '../AuthPanel'
 import TextInput from '$shared/components/TextInput'
 import Actions from '../Actions'
