+++ app/src/auth/utils/userAuthenticated.js
